+++ app/views/comments/_comment.html.erb
@@ -1,8 +1,6 @@
-<div class="comments_div">
-  <%= comments.each do |comment| %>
+
     <div class="single_comment_div">
     <p><strong><%= comment.author.name %>:</strong></p>
     <p><%= comment.text %></p>
     </div>
-  <% end %>
-</div>
+
