+++ app/views/comments/_comment_frm.html.erb
@@ -1,0 +1,9 @@
+<%= form_with model: Comment, url: user_post_comments_path(@user, @post), class: "comment_frm" do |f| %>
+    <div>
+      <%= f.text_field :text, class: "comment_text", placeholder: "Add new comment" %>
+    </div>
+  
+    <div>
+      <%= f.submit "Add comment", class: "button" %><br>
+    </div>
+<% end %>
