+++ config/routes.rb
@@ -5,7 +5,7 @@ Rails.application.routes.draw do
   root "users#index"
 
   resources :users, only: [:index, :show] do
-    resources :posts, only: [:index, :show, :new, :create] do
+    resources :posts, only: [:index, :new, :create, :show] do
       resources :comments, only: [:create]
       resources :likes, only: [:create]
     end
